markdown: Reduce mentions inside blockquotes to silent-mentions.
On the backend, we extend the BlockQuoteProcessor's clean function that
just removes '>' from the start of each line to convert each mention to
have the silent mention syntax, before UserMentionPattern is invoked.

The frontend, however, has an edge case where if you are mentioned in some
message and you quote it while having mentioned yourself above the quoted
message, you wouldn't see the red highlight till we get the final rendered
message from the backend.

Fixes #8025.